refactor(cli): move init command to app/cli/init.py
Cuts init_cmd and HELLO_CONDUIT_YAML out of app/main.py and into
app/cli/init.py. The decorator runs against the shared app from
app.cli.main during cli.main's side-effect imports, so init still
appears at the same position in --help (byte-identical to baseline).
